レスポンスをjson形式に変更。
