+++ gradle.properties
@@ -1,4 +1,5 @@
 ktor_version=1.6.7
 kotlin_version=1.6.10
 logback_version=1.2.3
+jackson_version=1.6.7
 kotlin.code.style=official
