+++ src/main/kotlin/com/example/controller/GreetController.kt
@@ -1,14 +1,17 @@
 package com.example.controller
 
+import com.example.domain.invoice.Invoice
 import io.ktor.application.*
 import io.ktor.response.*
 import io.ktor.routing.*
+import java.time.LocalDate
 
 fun Route.greetController() {
 
     route("/greet") {
         get ("/hello") {
-            call.respondText { "Hello!" }
+            val invoice = Invoice.create(1, 1000, LocalDate.parse("2022-01-10"))
+            call.respond(invoice)
         }
     }
