--- src/main/kotlin/com/example/controller/Routing.kt
@@ -1,11 +1,0 @@
-package com.example.controller
-
-import io.ktor.routing.*
-import io.ktor.application.*
-
-fun Application.Routing() {
-
-    routing {
-        greetController()
-    }
-}
