+++ src/main/kotlin/com/example/infrastructure/Jackson.kt
@@ -1,0 +1,18 @@
+package com.example.infrastructure
+
+import com.fasterxml.jackson.databind.SerializationFeature
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import io.ktor.application.*
+import io.ktor.features.*
+import io.ktor.jackson.*
+
+fun Application.jackson() {
+    install(ContentNegotiation) {
+        jackson {
+            configure(SerializationFeature.INDENT_OUTPUT, true)
+            registerModule(JavaTimeModule())
+            disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
+        }
+    }
+
+}
