+++ src/main/kotlin/com/example/infrastructure/Routing.kt
@@ -1,0 +1,12 @@
+package com.example.infrastructure
+
+import com.example.controller.greetController
+import io.ktor.routing.*
+import io.ktor.application.*
+
+fun Application.routing() {
+
+    routing {
+        greetController()
+    }
+}
